REFACTOR: (SPOT-166) Cache 패키지를 생성한다
