+++ src/main/java/com/meetup/server/global/cache/LoggingRedisCacheManager.java
@@ -1,6 +1,5 @@
-package com.meetup.server.global.support;
+package com.meetup.server.global.cache;
 
-import com.meetup.server.global.util.LoggingCache;
 import org.springframework.cache.Cache;
 import org.springframework.data.redis.cache.RedisCacheConfiguration;
 import org.springframework.data.redis.cache.RedisCacheManager;
